Redirects to `index.html`
If the `index` option is enabled and the request path corresponds to a
directory that contains a `index.html` file we need to render the
`index.html` file. Previously, the server just returned the file itself,
but this could cause issues if `index.html` contained some relative
links (e.g. to stylesheets). Now the server *redirects* to the file,
so that all the relative links work correctly.